requirements-e2e.txt
-r requirements-dev.txt
playwright>=1.40
pytest-playwright>=0.4
pytest-base-url>=2.1
requests>=2.31
